CreateFoodItem() finished
I started working more on foodItem. Clicking on a food item in the search results now creates a food scriptable object and then I made a method that will print out all of the nutrients contained in the item.

The way it can print this all out is because it contains a dictionary where:
<Key = nutrientID, Value = (value, unit) >

and then we just look up all the keys in another dictionary with
<Key = nutrientID, Value = Nutrient Scriptable Object>
which contains all the information related to the object

Co-Authored-By: leeyumsk <123519050+leeyumsk@users.noreply.github.com>