Adding extended nutrition information to the food selection info menu
This commit includes changes to the Food Selection Info Menu scene.

Slightly changed how the margins are calculated for the food name at the top of the screen in FoodSelectionInfoMenuInitializer.cs

Added a Scroll View which populates with Nutrition Label Entry prefabs, displaying the nutrient name and its value for each nutrient in the current food items nutrient dictionary.
(see NutrientDisplayPool.cs)